[TWEAK] Move attribute slicer into its own module
