+++ lib/woocommerce_api/resources/v2/variation.rb
@@ -4,6 +4,8 @@ require "woocommerce_api/resources/legacy/dimensions"
 module WoocommerceAPI
   module V2
     class Variation < Resource
+      include WoocommerceAPI::AttributeSlicer
+
       def load(attributes)
         # Rename restricted attributes
         if attributes['attributes']
@@ -20,17 +22,8 @@ module WoocommerceAPI
 
         variant_attributes.delete('image') unless options[:images]
         variant_attributes['backorders'] = nil if variant_attributes['backorders'].blank?
-        
-        case options[:sync_type]&.to_sym
-        when :price
-          variant_attributes.slice(:id, :regular_price, :sale_price)
-        when :stock_level
-          variant_attributes.slice(:id, :stock_quantity, :in_stock)
-        when :price_and_stock_level
-          variant_attributes.slice(:id, :regular_price, :sale_price, :stock_quantity, :in_stock)
-        else
-          variant_attributes
-        end
+
+        slice_by_sync_type(options[:sync_type], variant_attributes)
       end
 
       attribute :id, Integer
